Refactor: move StartGame() to GameLauncher object
Move some code for launching the game from qmldownloader.cpp to a new
file gamelauncher.cpp.
