+++ win.cpp
@@ -338,6 +338,7 @@ QString startGame(const QString& commandLine, bool failIfWindowsAdmin)
     }
 
     // Relaunch the updater without elevation so that the graphics preference can be set.
+    // TODO: propagate connect URL in this case
     std::wstring program = QCoreApplication::applicationFilePath().toStdWString();
     std::wstring args = L"--internalcommand playnow";
     qDebug() << "restarting de-elevated: program =" << program << "args =" << args;
